@@ -22,73 +22,73 @@
   },
   "dependencies": {
     "@4chain-ag/react-configuration": "^1.0.4",
-    "@bsv/sdk": "^1.0.37",
-    "@bsv/spv-wallet-js-client": "^1.0.0-beta.25",
+    "@bsv/sdk": "^1.2.11",
+    "@bsv/spv-wallet-js-client": "^1.0.0-beta.27",
     "@estruyf/github-actions-reporter": "^1.9.2",
-    "@heroicons/react": "^2.1.3",
-    "@hookform/resolvers": "^3.9.0",
-    "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-avatar": "^1.0.4",
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
+    "@heroicons/react": "^2.2.0",
+    "@hookform/resolvers": "^3.9.1",
+    "@radix-ui/react-accordion": "^1.2.1",
+    "@radix-ui/react-avatar": "^1.1.1",
+    "@radix-ui/react-dialog": "^1.1.2",
+    "@radix-ui/react-dropdown-menu": "^2.1.2",
     "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-popover": "^1.0.7",
-    "@radix-ui/react-radio-group": "^1.1.3",
-    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-popover": "^1.1.2",
+    "@radix-ui/react-radio-group": "^1.2.1",
+    "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-tooltip": "^1.1.2",
-    "@tanstack/react-query": "^5.37.1",
-    "@tanstack/react-router": "^1.33.4",
-    "@tanstack/react-table": "^8.17.3",
+    "@radix-ui/react-tabs": "^1.1.1",
+    "@radix-ui/react-tooltip": "^1.1.4",
+    "@tanstack/react-query": "^5.62.3",
+    "@tanstack/react-router": "^1.88.0",
+    "@tanstack/react-table": "^8.20.5",
     "@types/eslint__js": "^8.42.3",
-    "class-variance-authority": "^0.7.0",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "date-fns": "^3.6.0",
-    "lucide-react": "^0.436.0",
-    "next-themes": "^0.3.0",
-    "pino": "^9.1.0",
-    "pino-pretty": "^11.0.0",
-    "react": "^18.2.0",
-    "react-day-picker": "^8.10.1",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.52.2",
+    "date-fns": "^4.1.0",
+    "lucide-react": "^0.468.0",
+    "next-themes": "^0.4.4",
+    "pino": "^9.5.0",
+    "react": "^19.0.0",
+    "react-day-picker": "^9.4.2",
+    "react-dom": "^19.0.0",
+    "react-hook-form": "^7.54.0",
     "react-json-view": "^1.21.3",
-    "sonner": "^1.4.41",
-    "tailwind-merge": "^2.5.2",
+    "sonner": "^1.7.1",
+    "tailwind-merge": "^2.5.5",
     "tailwind-scrollbar-hide": "^1.1.7",
     "tailwindcss-animate": "^1.0.7",
-    "use-debounce": "^10.0.3",
-    "zod": "^3.23.8"
+    "use-debounce": "^10.0.4",
+    "vite-plugin-node-polyfills": "^0.22.0",
+    "zod": "^3.24.0"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.0.1",
-    "@eslint/js": "^9.3.0",
-    "@playwright/test": "^1.47.2",
-    "@tanstack/eslint-plugin-query": "^5.35.6",
-    "@tanstack/router-devtools": "^1.33.4",
-    "@tanstack/router-vite-plugin": "^1.32.17",
-    "@types/node": "^20.12.12",
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
-    "@typescript-eslint/parser": "^7.2.0",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "@vitest/coverage-v8": "^2.0.5",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^9.8.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-react": "^7.35.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "globals": "^15.3.0",
-    "postcss": "^8.4.38",
+    "@eslint/compat": "^1.2.4",
+    "@eslint/js": "^9.16.0",
+    "@playwright/test": "^1.49.1",
+    "@tanstack/eslint-plugin-query": "^5.62.1",
+    "@tanstack/router-devtools": "^1.88.0",
+    "@tanstack/router-vite-plugin": "^1.87.13",
+    "@types/node": "^22.10.1",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
+    "@typescript-eslint/parser": "^8.18.0",
+    "@vitejs/plugin-react-swc": "^3.7.2",
+    "@vitest/coverage-v8": "^2.1.8",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.16.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.16",
+    "globals": "^15.13.0",
+    "postcss": "^8.4.49",
     "prettier": "3.2.5",
-    "tailwindcss": "^3.4.3",
-    "typescript": "^5.2.2",
-    "typescript-eslint": "^7.9.0",
-    "vite": "^5.2.0",
+    "tailwindcss": "^3.4.16",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.0",
+    "vite": "^6.0.3",
     "vite-plugin-eslint": "^1.8.1",
-    "vitest": "^2.0.5"
+    "vitest": "^2.1.8"
   },
   "resolutions": {
     "strip-ansi": "6.0.1"
