@@ -5,7 +5,6 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-  Form,
   FormControl,
   FormField,
   FormItem,
@@ -21,6 +20,7 @@ import {
   SelectValue,
   Toaster,
 } from '@/components';
+import { FormProvider as Form } from 'react-hook-form';
 import { Role, TRole, LoginType, useAuth, useSpvWalletClient } from '@/contexts';
 
 import { createClient, errorWrapper, getShortXprv } from '@/utils';
@@ -61,6 +61,7 @@ export function LoginForm() {
 
   const { config } = useConfig();
   const { configureServerUrl = false } = config;
+  const isServerUrlInitialized = useRef(false);
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
@@ -68,7 +69,7 @@ export function LoginForm() {
       role: Role.Admin,
       type: LoginType.Xprv,
       key: '',
-      serverUrl: serverUrl,
+      serverUrl: '',
     },
   });
 
@@ -77,6 +78,13 @@ export function LoginForm() {
 
   const currentRole = form.getValues('role');
   const currentType = form.getValues('type');
+
+  useEffect(() => {
+    if (serverUrl && !isServerUrlInitialized.current) {
+      form.reset((prev) => ({ ...prev, serverUrl }));
+      isServerUrlInitialized.current = true;
+    }
+  }, [serverUrl, form.reset]);
 
   useEffect(() => {
     inputRef.current?.focus();
