@@ -1,12 +1,12 @@
-/* prettier-ignore-start */
-
 /* eslint-disable */
 
 // @ts-nocheck
 
 // noinspection JSUnusedGlobalSymbols
 
-// This file is auto-generated by TanStack Router
+// This file was automatically generated by TanStack Router.
+// You should NOT make any changes in this file as it will be overwritten.
+// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
 
 import { createFileRoute } from '@tanstack/react-router'
 
@@ -36,26 +36,31 @@ const AboutLazyImport = createFileRoute('/about')()
 // Create/Update Routes
 
 const UserRoute = UserImport.update({
+  id: '/user',
   path: '/user',
   getParentRoute: () => rootRoute,
 } as any)
 
 const AdminRoute = AdminImport.update({
+  id: '/admin',
   path: '/admin',
   getParentRoute: () => rootRoute,
 } as any)
 
 const AboutLazyRoute = AboutLazyImport.update({
+  id: '/about',
   path: '/about',
   getParentRoute: () => rootRoute,
 } as any).lazy(() => import('./about.lazy').then((d) => d.Route))
 
 const LoginRoute = LoginImport.update({
+  id: '/login',
   path: '/login',
   getParentRoute: () => rootRoute,
 } as any)
 
 const IndexRoute = IndexImport.update({
+  id: '/',
   path: '/',
   getParentRoute: () => rootRoute,
 } as any)
@@ -71,46 +76,55 @@ const AdminAdminRoute = AdminAdminImport.update({
 } as any)
 
 const UserUserXpubRoute = UserUserXpubImport.update({
+  id: '/xpub',
   path: '/xpub',
   getParentRoute: () => UserUserRoute,
 } as any)
 
 const UserUserTransactionsRoute = UserUserTransactionsImport.update({
+  id: '/transactions',
   path: '/transactions',
   getParentRoute: () => UserUserRoute,
 } as any)
 
 const UserUserAccessKeysRoute = UserUserAccessKeysImport.update({
+  id: '/access-keys',
   path: '/access-keys',
   getParentRoute: () => UserUserRoute,
 } as any)
 
 const AdminAdminXpubRoute = AdminAdminXpubImport.update({
+  id: '/xpub',
   path: '/xpub',
   getParentRoute: () => AdminAdminRoute,
 } as any)
 
 const AdminAdminWebhooksRoute = AdminAdminWebhooksImport.update({
+  id: '/webhooks',
   path: '/webhooks',
   getParentRoute: () => AdminAdminRoute,
 } as any)
 
 const AdminAdminTransactionsRoute = AdminAdminTransactionsImport.update({
+  id: '/transactions',
   path: '/transactions',
   getParentRoute: () => AdminAdminRoute,
 } as any)
 
 const AdminAdminPaymailsRoute = AdminAdminPaymailsImport.update({
+  id: '/paymails',
   path: '/paymails',
   getParentRoute: () => AdminAdminRoute,
 } as any)
 
 const AdminAdminContactsRoute = AdminAdminContactsImport.update({
+  id: '/contacts',
   path: '/contacts',
   getParentRoute: () => AdminAdminRoute,
 } as any)
 
 const AdminAdminAccessKeysRoute = AdminAdminAccessKeysImport.update({
+  id: '/access-keys',
   path: '/access-keys',
   getParentRoute: () => AdminAdminRoute,
 } as any)
@@ -236,27 +250,283 @@ declare module '@tanstack/react-router' {
 
 // Create and export the route tree
 
-export const routeTree = rootRoute.addChildren({
-  IndexRoute,
-  LoginRoute,
-  AboutLazyRoute,
-  AdminRoute: AdminRoute.addChildren({
-    AdminAdminRoute: AdminAdminRoute.addChildren({
-      AdminAdminAccessKeysRoute,
-      AdminAdminContactsRoute,
-      AdminAdminPaymailsRoute,
-      AdminAdminTransactionsRoute,
-      AdminAdminWebhooksRoute,
-      AdminAdminXpubRoute,
-    }),
-  }),
-  UserRoute: UserRoute.addChildren({
-    UserUserRoute: UserUserRoute.addChildren({
-      UserUserAccessKeysRoute,
-      UserUserTransactionsRoute,
-      UserUserXpubRoute,
-    }),
-  }),
-})
+interface AdminAdminRouteChildren {
+  AdminAdminAccessKeysRoute: typeof AdminAdminAccessKeysRoute
+  AdminAdminContactsRoute: typeof AdminAdminContactsRoute
+  AdminAdminPaymailsRoute: typeof AdminAdminPaymailsRoute
+  AdminAdminTransactionsRoute: typeof AdminAdminTransactionsRoute
+  AdminAdminWebhooksRoute: typeof AdminAdminWebhooksRoute
+  AdminAdminXpubRoute: typeof AdminAdminXpubRoute
+}
 
-/* prettier-ignore-end */
+const AdminAdminRouteChildren: AdminAdminRouteChildren = {
+  AdminAdminAccessKeysRoute: AdminAdminAccessKeysRoute,
+  AdminAdminContactsRoute: AdminAdminContactsRoute,
+  AdminAdminPaymailsRoute: AdminAdminPaymailsRoute,
+  AdminAdminTransactionsRoute: AdminAdminTransactionsRoute,
+  AdminAdminWebhooksRoute: AdminAdminWebhooksRoute,
+  AdminAdminXpubRoute: AdminAdminXpubRoute,
+}
+
+const AdminAdminRouteWithChildren = AdminAdminRoute._addFileChildren(
+  AdminAdminRouteChildren,
+)
+
+interface AdminRouteChildren {
+  AdminAdminRoute: typeof AdminAdminRouteWithChildren
+}
+
+const AdminRouteChildren: AdminRouteChildren = {
+  AdminAdminRoute: AdminAdminRouteWithChildren,
+}
+
+const AdminRouteWithChildren = AdminRoute._addFileChildren(AdminRouteChildren)
+
+interface UserUserRouteChildren {
+  UserUserAccessKeysRoute: typeof UserUserAccessKeysRoute
+  UserUserTransactionsRoute: typeof UserUserTransactionsRoute
+  UserUserXpubRoute: typeof UserUserXpubRoute
+}
+
+const UserUserRouteChildren: UserUserRouteChildren = {
+  UserUserAccessKeysRoute: UserUserAccessKeysRoute,
+  UserUserTransactionsRoute: UserUserTransactionsRoute,
+  UserUserXpubRoute: UserUserXpubRoute,
+}
+
+const UserUserRouteWithChildren = UserUserRoute._addFileChildren(
+  UserUserRouteChildren,
+)
+
+interface UserRouteChildren {
+  UserUserRoute: typeof UserUserRouteWithChildren
+}
+
+const UserRouteChildren: UserRouteChildren = {
+  UserUserRoute: UserUserRouteWithChildren,
+}
+
+const UserRouteWithChildren = UserRoute._addFileChildren(UserRouteChildren)
+
+export interface FileRoutesByFullPath {
+  '/': typeof IndexRoute
+  '/login': typeof LoginRoute
+  '/about': typeof AboutLazyRoute
+  '/admin': typeof AdminAdminRouteWithChildren
+  '/user': typeof UserUserRouteWithChildren
+  '/admin/access-keys': typeof AdminAdminAccessKeysRoute
+  '/admin/contacts': typeof AdminAdminContactsRoute
+  '/admin/paymails': typeof AdminAdminPaymailsRoute
+  '/admin/transactions': typeof AdminAdminTransactionsRoute
+  '/admin/webhooks': typeof AdminAdminWebhooksRoute
+  '/admin/xpub': typeof AdminAdminXpubRoute
+  '/user/access-keys': typeof UserUserAccessKeysRoute
+  '/user/transactions': typeof UserUserTransactionsRoute
+  '/user/xpub': typeof UserUserXpubRoute
+}
+
+export interface FileRoutesByTo {
+  '/': typeof IndexRoute
+  '/login': typeof LoginRoute
+  '/about': typeof AboutLazyRoute
+  '/admin': typeof AdminAdminRouteWithChildren
+  '/user': typeof UserUserRouteWithChildren
+  '/admin/access-keys': typeof AdminAdminAccessKeysRoute
+  '/admin/contacts': typeof AdminAdminContactsRoute
+  '/admin/paymails': typeof AdminAdminPaymailsRoute
+  '/admin/transactions': typeof AdminAdminTransactionsRoute
+  '/admin/webhooks': typeof AdminAdminWebhooksRoute
+  '/admin/xpub': typeof AdminAdminXpubRoute
+  '/user/access-keys': typeof UserUserAccessKeysRoute
+  '/user/transactions': typeof UserUserTransactionsRoute
+  '/user/xpub': typeof UserUserXpubRoute
+}
+
+export interface FileRoutesById {
+  __root__: typeof rootRoute
+  '/': typeof IndexRoute
+  '/login': typeof LoginRoute
+  '/about': typeof AboutLazyRoute
+  '/admin': typeof AdminRouteWithChildren
+  '/admin/_admin': typeof AdminAdminRouteWithChildren
+  '/user': typeof UserRouteWithChildren
+  '/user/_user': typeof UserUserRouteWithChildren
+  '/admin/_admin/access-keys': typeof AdminAdminAccessKeysRoute
+  '/admin/_admin/contacts': typeof AdminAdminContactsRoute
+  '/admin/_admin/paymails': typeof AdminAdminPaymailsRoute
+  '/admin/_admin/transactions': typeof AdminAdminTransactionsRoute
+  '/admin/_admin/webhooks': typeof AdminAdminWebhooksRoute
+  '/admin/_admin/xpub': typeof AdminAdminXpubRoute
+  '/user/_user/access-keys': typeof UserUserAccessKeysRoute
+  '/user/_user/transactions': typeof UserUserTransactionsRoute
+  '/user/_user/xpub': typeof UserUserXpubRoute
+}
+
+export interface FileRouteTypes {
+  fileRoutesByFullPath: FileRoutesByFullPath
+  fullPaths:
+    | '/'
+    | '/login'
+    | '/about'
+    | '/admin'
+    | '/user'
+    | '/admin/access-keys'
+    | '/admin/contacts'
+    | '/admin/paymails'
+    | '/admin/transactions'
+    | '/admin/webhooks'
+    | '/admin/xpub'
+    | '/user/access-keys'
+    | '/user/transactions'
+    | '/user/xpub'
+  fileRoutesByTo: FileRoutesByTo
+  to:
+    | '/'
+    | '/login'
+    | '/about'
+    | '/admin'
+    | '/user'
+    | '/admin/access-keys'
+    | '/admin/contacts'
+    | '/admin/paymails'
+    | '/admin/transactions'
+    | '/admin/webhooks'
+    | '/admin/xpub'
+    | '/user/access-keys'
+    | '/user/transactions'
+    | '/user/xpub'
+  id:
+    | '__root__'
+    | '/'
+    | '/login'
+    | '/about'
+    | '/admin'
+    | '/admin/_admin'
+    | '/user'
+    | '/user/_user'
+    | '/admin/_admin/access-keys'
+    | '/admin/_admin/contacts'
+    | '/admin/_admin/paymails'
+    | '/admin/_admin/transactions'
+    | '/admin/_admin/webhooks'
+    | '/admin/_admin/xpub'
+    | '/user/_user/access-keys'
+    | '/user/_user/transactions'
+    | '/user/_user/xpub'
+  fileRoutesById: FileRoutesById
+}
+
+export interface RootRouteChildren {
+  IndexRoute: typeof IndexRoute
+  LoginRoute: typeof LoginRoute
+  AboutLazyRoute: typeof AboutLazyRoute
+  AdminRoute: typeof AdminRouteWithChildren
+  UserRoute: typeof UserRouteWithChildren
+}
+
+const rootRouteChildren: RootRouteChildren = {
+  IndexRoute: IndexRoute,
+  LoginRoute: LoginRoute,
+  AboutLazyRoute: AboutLazyRoute,
+  AdminRoute: AdminRouteWithChildren,
+  UserRoute: UserRouteWithChildren,
+}
+
+export const routeTree = rootRoute
+  ._addFileChildren(rootRouteChildren)
+  ._addFileTypes<FileRouteTypes>()
+
+/* ROUTE_MANIFEST_START
+{
+  "routes": {
+    "__root__": {
+      "filePath": "__root.tsx",
+      "children": [
+        "/",
+        "/login",
+        "/about",
+        "/admin",
+        "/user"
+      ]
+    },
+    "/": {
+      "filePath": "index.tsx"
+    },
+    "/login": {
+      "filePath": "login.tsx"
+    },
+    "/about": {
+      "filePath": "about.lazy.tsx"
+    },
+    "/admin": {
+      "filePath": "admin",
+      "children": [
+        "/admin/_admin"
+      ]
+    },
+    "/admin/_admin": {
+      "filePath": "admin/_admin.tsx",
+      "parent": "/admin",
+      "children": [
+        "/admin/_admin/access-keys",
+        "/admin/_admin/contacts",
+        "/admin/_admin/paymails",
+        "/admin/_admin/transactions",
+        "/admin/_admin/webhooks",
+        "/admin/_admin/xpub"
+      ]
+    },
+    "/user": {
+      "filePath": "user",
+      "children": [
+        "/user/_user"
+      ]
+    },
+    "/user/_user": {
+      "filePath": "user/_user.tsx",
+      "parent": "/user",
+      "children": [
+        "/user/_user/access-keys",
+        "/user/_user/transactions",
+        "/user/_user/xpub"
+      ]
+    },
+    "/admin/_admin/access-keys": {
+      "filePath": "admin/_admin.access-keys.tsx",
+      "parent": "/admin/_admin"
+    },
+    "/admin/_admin/contacts": {
+      "filePath": "admin/_admin.contacts.tsx",
+      "parent": "/admin/_admin"
+    },
+    "/admin/_admin/paymails": {
+      "filePath": "admin/_admin.paymails.tsx",
+      "parent": "/admin/_admin"
+    },
+    "/admin/_admin/transactions": {
+      "filePath": "admin/_admin.transactions.tsx",
+      "parent": "/admin/_admin"
+    },
+    "/admin/_admin/webhooks": {
+      "filePath": "admin/_admin.webhooks.tsx",
+      "parent": "/admin/_admin"
+    },
+    "/admin/_admin/xpub": {
+      "filePath": "admin/_admin.xpub.tsx",
+      "parent": "/admin/_admin"
+    },
+    "/user/_user/access-keys": {
+      "filePath": "user/_user.access-keys.tsx",
+      "parent": "/user/_user"
+    },
+    "/user/_user/transactions": {
+      "filePath": "user/_user.transactions.tsx",
+      "parent": "/user/_user"
+    },
+    "/user/_user/xpub": {
+      "filePath": "user/_user.xpub.tsx",
+      "parent": "/user/_user"
+    }
+  }
+}
+ROUTE_MANIFEST_END */
